moving all packages from core.* to core.auth_service.* and changing data type of primary key in user_cred
