+++ src/main/java/com/hangout/core/auth_service/UsersApplication.java
@@ -1,0 +1,20 @@
+package com.hangout.core.auth_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.info.Info;
+import io.swagger.v3.oas.annotations.servers.Server;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+@OpenAPIDefinition(info = @Info(title = "Hangout Auth API", version = "v1", description = "This API handles user Authentication and Authorization for all services in Hangout"), servers = @Server(url = "http://localhost:5011"))
+public class UsersApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UsersApplication.class, args);
+	}
+
+}
