+++ src/main/java/com/hangout/core/auth_service/config/BcryptConfig.java
@@ -1,0 +1,18 @@
+package com.hangout.core.auth_service.config;
+
+import java.security.SecureRandom;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder.BCryptVersion;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+public class BcryptConfig {
+    @Bean
+    PasswordEncoder passwordEncoder() {
+        SecureRandom random = new SecureRandom();
+        return new BCryptPasswordEncoder(BCryptVersion.$2B, 11, new SecureRandom(random.generateSeed(20)));
+    }
+}
