+++ src/main/java/com/hangout/core/auth_service/config/ObserverConfig.java
@@ -1,0 +1,15 @@
+package com.hangout.core.auth_service.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import io.micrometer.observation.ObservationRegistry;
+import io.micrometer.observation.aop.ObservedAspect;
+
+@Configuration(proxyBeanMethods = false)
+public class ObserverConfig {
+    @Bean
+    ObservedAspect observedAspect(ObservationRegistry observationRegistry) {
+        return new ObservedAspect(observationRegistry);
+    }
+}
