+++ src/main/java/com/hangout/core/auth_service/dto/ExistingUser.java
@@ -1,0 +1,5 @@
+package com.hangout.core.auth_service.dto;
+
+public record ExistingUser(String username, String password) {
+
+}
