+++ src/main/java/com/hangout/core/auth_service/dto/NewUser.java
@@ -1,0 +1,5 @@
+package com.hangout.core.auth_service.dto;
+
+public record NewUser(String username, String email, String password) {
+
+}
