+++ src/main/java/com/hangout/core/auth_service/entity/Roles.java
@@ -1,0 +1,5 @@
+package com.hangout.core.auth_service.entity;
+
+public enum Roles {
+	USER, ADMIN
+}
