+++ src/main/java/com/hangout/core/auth_service/exceptions/EmailOrPasswordWrong.java
@@ -1,0 +1,19 @@
+package com.hangout.core.auth_service.exceptions;
+
+public class EmailOrPasswordWrong extends RuntimeException {
+	private static final long serialVersionUID = -5369312191560192519L;
+	private String message;
+
+	public EmailOrPasswordWrong() {
+		super();
+	}
+
+	public EmailOrPasswordWrong(String message) {
+		this.message = message;
+	}
+
+	public String getMessage() {
+		return message;
+	}
+
+}
