+++ src/main/java/com/hangout/core/auth_service/exceptions/handlers/GlobalExceptionHandler.java
@@ -1,0 +1,20 @@
+package com.hangout.core.auth_service.exceptions.handlers;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ProblemDetail;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import com.hangout.core.auth_service.exceptions.UserNotFoundException;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
+	@ExceptionHandler(UserNotFoundException.class)
+	public ProblemDetail UserNotValidHandler(UserNotFoundException ex) {
+		ProblemDetail problem = ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, ex.getMessage());
+		problem.setTitle("Given user/s not found");
+		return problem;
+	}
+
+}
