--- src/main/java/com/hangout/core/controller/UserController.java
@@ -1,40 +1,0 @@
-package com.hangout.core.controller;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.hangout.core.service.UserDetailsServiceImpl;
-
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.extern.slf4j.Slf4j;
-
-@RestController
-@RequestMapping("/v1/user")
-@Tag(name = "Protected Endpoints")
-@Slf4j
-public class UserController {
-	@Autowired
-	private UserDetailsServiceImpl userDetailsService;
-
-	@DeleteMapping
-	public ResponseEntity<String> deleteUser() {
-		try {
-			String userName = getAuthenticatedUser().getName();
-			this.userDetailsService.deleteUser(userName);
-			return new ResponseEntity<>("User with username: " + userName + " deleted", HttpStatus.OK);
-		} catch (Exception e) {
-			log.error("Exception: {}", e.getCause());
-			return new ResponseEntity<>("User could not be deleted", HttpStatus.BAD_REQUEST);
-		}
-	}
-
-	private Authentication getAuthenticatedUser() {
-		return SecurityContextHolder.getContext().getAuthentication();
-	}
-}
