--- src/main/java/com/hangout/core/dto/ExistingUser.java
@@ -1,5 +1,0 @@
-package com.hangout.core.dto;
-
-public record ExistingUser(String username, String password) {
-
-}
