--- src/main/java/com/hangout/core/dto/NewUser.java
@@ -1,5 +1,0 @@
-package com.hangout.core.dto;
-
-public record NewUser(String username, String email, String password) {
-
-}
