--- src/main/java/com/hangout/core/entity/Gender.java
@@ -1,5 +1,0 @@
-package com.hangout.core.entity;
-
-public enum Gender {
-    male, female, other
-}
