--- src/main/java/com/hangout/core/entity/Roles.java
@@ -1,5 +1,0 @@
-package com.hangout.core.entity;
-
-public enum Roles {
-	USER, ADMIN
-}
