--- src/main/java/com/hangout/core/exceptions/handlers/GlobalExceptionHandler.java
@@ -1,20 +1,0 @@
-package com.hangout.core.exceptions.handlers;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ProblemDetail;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-import com.hangout.core.exceptions.UserNotFoundException;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
-	@ExceptionHandler(UserNotFoundException.class)
-	public ProblemDetail UserNotValidHandler(UserNotFoundException ex) {
-		ProblemDetail problem = ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, ex.getMessage());
-		problem.setTitle("Given user/s not found");
-		return problem;
-	}
-
-}
