+++ src/test/java/com/hangout/core/auth_service/service/UserServiceTest.java
@@ -1,0 +1,8 @@
+package com.hangout.core.auth_service.service;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class UserServiceTest {
+
+}
