--- src/test/java/com/hangout/core/service/UserServiceTest.java
@@ -1,8 +1,0 @@
-package com.hangout.core.service;
-
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class UserServiceTest {
-
-}
